symtbl: capture function parameter types, env and locs
A caveat here is the use of the function keyword with built-in
pattern matching : the implicit argument is named `param` past the lambda
pass, but does not appear in typedtree. Name conflict might appear with
lookups, so a revamp of the structure used to represent the symbol table
will be required at some point.

For now, implicit arguments wil be skipped.