Utilities to determine memory region of an address
It is now possible to determine stack and heap addresses at runtime.
If neither, the data sections are checked.

Those utilities are introduced because OCaml pointers might be stored on
the stack, and an extra indirection might be needed to be able to
display/pretty print that value.

lldb does provide a method through its API to
get a list of memory regions of the process called GetMemoryRegions,
but for 3.9 only (see http://reviews.llvm.org/D20565)

Until the lldb OCaml plugin is integrated into lldb 3.9, so in the
meantime, we read and parse the maps files associated to a process in
the /proc pseudo filesystem to check to which memory region a specific address of a process belongs to.

It is obviously not portable, so the main target is assumed to be a Linux x86 64 bits system.