All OCaml custom commands return strings
The separate printlining of OCaml commands was a pain because not
getting along with the standard commands outputs when running them all
together in batch mode.
Now all commands, custom or not, have their output in a
CommandReturnObject.
The DoExecute method in OCamlCommand running the ocaml command
callback is done regardless anyway.